+++ packages/orq-rc/src/models/operations/createprompt.ts
@@ -5950,7 +5950,7 @@ export const CreatePromptResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-02-02T16:01:07.474Z",
+    "2025-02-03T06:31:29.908Z",
   ).transform(v => new Date(v)),
   type: CreatePromptPromptsType$inboundSchema,
   versions: z.array(z.lazy(() => CreatePromptVersions$inboundSchema)),
@@ -5998,7 +5998,7 @@ export const CreatePromptResponseBody$outboundSchema: z.ZodType<
   promptConfig: z.lazy(() => CreatePromptPromptsPromptConfig$outboundSchema),
   metadata: z.lazy(() => CreatePromptPromptsMetadata$outboundSchema),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-02-02T16:01:07.474Z"))
+  updated: z.date().default(() => new Date("2025-02-03T06:31:29.908Z"))
     .transform(v => v.toISOString()),
   type: CreatePromptPromptsType$outboundSchema,
   versions: z.array(z.lazy(() => CreatePromptVersions$outboundSchema)),